fix: fix typing issue in schema. fix n+1 queries for project. fix typing mismatch in openapi

+++ backend/ops_api/ops/schemas/projects.py
@@ -1,4 +1,5 @@
 from datetime import date, datetime
+from decimal import Decimal
 from typing import Optional
 
 from marshmallow import Schema, fields, pre_dump
@@ -150,7 +151,7 @@ class ProjectListResponse(Schema):
     fiscal_year_totals: Optional[dict] = fields.Dict(
         keys=fields.Int(), values=fields.Decimal(as_string=True), allow_none=True
     )
-    project_total: Optional[int] = fields.Decimal(allow_none=True)
+    project_total: Optional[Decimal] = fields.Decimal(allow_none=True)
     agreement_name_list: Optional[list[dict]] = fields.List(
         fields.Nested(AgreementNameListItem), dump_default=[], allow_none=True
     )
